requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
dnspython==2.3.0
idna==3.10
pymongo==4.7.3
python-dotenv==0.21.1
requests==2.31.0
urllib3==2.0.7
